Add token Factory
It can generate [unsecured
tokens](https://tools.ietf.org/html/rfc7519#section-6) using the "none"
algorithm. This can be achieved by not passing a signer when
instantiating the factory.

When creating unsecured tokens, the payload will intentionally have a
trailing dot since as per the RFC the signature is in this case an empty
string.