Add None signer
I was on the fence about this... By definition, using the 'none'
algorithm means the token will never pass verification, and the
signature should always be an empty string...

But allowing for the signer to be optional may lead to confusion and to
accidentally creating unsecured tokens if the user forgets to pass it to
the factory.

On the plus side, it lets me simplify the logic on the factory a bit.